requirements.txt
aniso8601==10.0.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
pytz==2025.2
six==1.17.0
Werkzeug==3.1.3
